requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.10.4
openai==1.59.3
python-dotenv==1.0.1
jinja2==3.1.5
python-multipart==0.0.20
readability-lxml==0.8.1
trafilatura==2.0.0
tldextract==5.1.3
